initial commit of ssl-cert-checker
high hopes for github actions doing linting and multi-arch build.

this will produce a docker image that can take a list of sites to connect to and check the expiration of their ssl certificates. the container will run a web server that has a /metrics endpoint in which prometheus can scrape for export.